Revert back to 0.13.0-SNAPSHOT
The 0.13.1-SNAPSHOT artifact version is used by the release/0.13 branch to prepare for the first patch of 0.13 (that is,
0.13.1). Utilizing the version here on the default branch will create packages that could interfere with that development line when
it comes into existence.

Additionally, other logic that attempts to find the latest version will return 0.13.1-SNAPSHOT as being newer than 0.13.0 (full
release), even though this isn't true. When used in the release branch, this relationship does hold -- development in preparation
for the first patch _is_ newer than the last release.